Split browser execution primitives
